+++ src/TestTrackingDiagrams.Extensions.StorageQueues/StorageQueueTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.StorageQueues;
+
+public enum StorageQueueTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}